Create BareRootProductGetterBase for FireworksWeb
The BareRootProductGetterBase has nearly all the functionality of
earlier BareRootProductGetter, except the logic to get the current
file in getIt(), that is now a pure virtual function, overridden and
implemented in BareRootProductGetter. FireworksWeb can now inherit
from BareRootProductGetterBase and implement only the currentFile().
